Add reusable upgrade audit intelligence surface
